requirements.txt

streamlit
pandas
numpy
openpyxl
pytesseract
Pillow
pyngrok
